Moving rate limit to get a taste of the changes
