+++ opendbc/car/controls/lateral.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass
 from opendbc.car import structs
 import numpy as np
 
+FRICTION_THRESHOLD = 0.3
 
 @dataclass
 class AngleSteeringLimits:
