+++ opendbc/car/toyota/carcontroller.py
@@ -1,6 +1,7 @@
 import math
 import numpy as np
-from opendbc.car import Bus, make_tester_present_msg, rate_limit, structs, ACCELERATION_DUE_TO_GRAVITY, DT_CTRL
+from opendbc.car import Bus, make_tester_present_msg, structs, ACCELERATION_DUE_TO_GRAVITY, DT_CTRL
+from opendbc.car.controls import rate_limit
 from opendbc.car.controls.lateral import apply_meas_steer_torque_limits, apply_std_steer_angle_limits, common_fault_avoidance
 from opendbc.car.can_definitions import CanData
 from opendbc.car.carlog import carlog
